Add basic unit tests for CustomerService

+++ src/main/java/com/replicantt/energyrs/service/CustomerService.java
@@ -8,6 +8,9 @@ import org.springframework.stereotype.Service;
 import com.replicantt.energyrs.repository.Customer;
 import com.replicantt.energyrs.repository.CustomerRepository;
 
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
 @Service
 public class CustomerService {
     private final CustomerRepository customerRepository;
@@ -18,15 +21,23 @@ public class CustomerService {
     public List<Customer> getAllCustomers() {
         return customerRepository.findAll();
     }
+    public Customer getCustomer(Long id) {
+        return customerRepository.findById(id).orElseThrow(() -> new RuntimeException("Customer not found"));
+    }
 
     public Customer addCustomer(Customer customer) {
+        log.debug("Customer saved: {}", customer);
         return customerRepository.save(customer);
     }
 
     public void deleteCustomer(Long id) {
-        customerRepository.deleteById(id);
+        if (!customerRepository.existsById(id)) {
+            throw new RuntimeException("Customer not found with ID: " + id);
+        } else { 
+            customerRepository.deleteById(id);
+        }
     }
-    public void updateCustomer(Long id, String name, String email, LocalDate birth, String phoneNumber, String address) {
+    public Customer updateCustomer(Long id, String name, String email, LocalDate birth, String phoneNumber, String address) {
         Customer customer = customerRepository.findById(id).orElseThrow(() -> new RuntimeException("Customer not found"));
         
         customer.setName(name);
@@ -35,6 +46,6 @@ public class CustomerService {
         customer.setPhoneNumber(phoneNumber);
         customer.setAddress(address);
 
-        customerRepository.save(customer);
+        return customerRepository.save(customer);
     }
 }
